# openapi.yaml
openapi: 3.1.0
info:
  title: Brain RH API
  version: 1.0.0
  description: |
    API de matching intelligent CV/Offre d'emploi avec analyse LLM.

    **Fonctionnalités:**
    - Parsing de CVs (PDF/DOCX) avec extraction LLM
    - Enrichissement d'offres avec génération must-have/nice-have
    - Matching avec filtrage, scoring et re-ranking LLM
    - Streaming temps-réel (SSE) pour traitements longs
    - Gestion de projets et entreprises
    - Export CSV/JSON des résultats

  contact:
    name: Brain RH API Support
    email: support@brainrh.com
servers:
  - url: http://localhost:8000/api/v1
    description: Serveur de développement
  - url: https://api.brainrh.com/api/v1
    description: Serveur de production
...
paths:
...
  /projects:
    get:
      tags: [Projets]
      summary: Lister tous les projets
      operationId: listProjects
      parameters:
        - name: enterprise_id
          in: query
          schema:
            type: string
          description: Filtrer par entreprise
        - name: status
          in: query
          schema:
            type: string
            enum: [actif, archive]
          description: Filtrer par status
      responses:
        '200':
          description: Liste des projets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Project'

    post:
      tags: [Projets]
      summary: Créer un projet
      operationId: createProject
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectInput'
      responses:
        '201':
          description: Projet créé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '400':
          $ref: '#/components/responses/BadRequest'
  /projects/{project_id}:
    get:
      tags: [Projets]
      summary: Récupérer un projet
      operationId: getProject
      parameters:
        - name: project_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Projet trouvé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags: [Projets]
      summary: Modifier un projet
      operationId: updateProject
      parameters:
        - name: project_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectInput'
      responses:
        '200':
          description: Projet modifié
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [Projets]
      summary: Supprimer un projet
      operationId: deleteProject
      parameters:
        - name: project_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Projet supprimé
        '404':
          $ref: '#/components/responses/NotFound'
  /projects/{project_id}/history:
    get:
      tags: [Projets]
      summary: Historique des matchings d'un projet
      operationId: getProjectHistory
      parameters:
        - name: project_id
          in: path
          required: true
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: Historique trouvé
          content:
            application/json:
              schema:
                type: object
                properties:
                  total:
                    type: integer
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/MatchingHistoryEntry'

  # ==================== ENTREPRISES ====================
...
components:
  schemas:
...
    MatchingHistoryEntry:
      type: object
      properties:
        matching_id:
          type: string
        timestamp:
          type: string
          format: date-time
        offre_titre:
          type: string
        total_cvs:
          type: integer
        top_score:
          type: number
          format: float

    # ==================== PROJETS ====================
    Project:
      type: object
      required:
        - id
        - nom
        - created_at
        - last_modified
      properties:
        id:
          type: string
        nom:
          type: string
        enterprise_id:
          type: string
        created_at:
          type: string
          format: date-time
        last_modified:
          type: string
          format: date-time
        status:
          type: string
          enum: [actif, archive]
          default: actif
    ProjectInput:
      type: object
      required:
        - nom
      properties:
        nom:
          type: string
        enterprise_id:
          type: string
        status:
          type: string
          enum: [actif, archive]

    # ==================== ENTREPRISES ====================
...
    APIError:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          description: Code d'erreur machine-readable
        message:
          type: string
          description: Message d'erreur human-readable
        details:
          type: object
          description: Détails additionnels
        status:
          type: integer
          description: Code HTTP
  responses:
    BadRequest:
      description: Requête invalide
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/APIError'
          example:
            code: INVALID_REQUEST
            message: "Le champ 'files' est requis"
            status: 400
    NotFound:
      description: Ressource non trouvée
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/APIError'
          example:
            code: NOT_FOUND
            message: "Ressource non trouvée"
            status: 404
...
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: Clé API pour authentification (à implémenter si nécessaire)

# Note: Pas de sécurité activée pour le moment (développement local)
# security:
#   - ApiKeyAuth: []